Fast Track: Split CLI from runner + validate BASE_REF
P2: extract the local CLI into cli.ts so run-guardrails.ts is a pure
runner with only type imports — no longer dragging the git provider and
registry into every importer (matches its "shared by CI and local"
docstring; keeps C2b's CI entry clean).

P1: validate BASE_REF before judging — reject an empty ref (which would
diff `...HEAD`, a vacuous empty pass) in the GitGuardrailContext ctor, and
assert the ref resolves to a commit so a typo fails loudly instead of
silently passing.
