+++ fast_track/package.json
@@ -12,8 +12,8 @@
         "format": "prettier --write --cache --log-level warn . && eslint . --fix",
         "static-checks": "npm run lint && npm run check",
         "test": "vitest run",
-        "guardrails": "tsx src/guardrails/run-guardrails.ts",
-        "list-guardrails": "tsx src/guardrails/run-guardrails.ts --list"
+        "guardrails": "tsx src/guardrails/cli.ts",
+        "list-guardrails": "tsx src/guardrails/cli.ts --list"
     },
     "dependencies": {
         "simple-git": "^3.36.0"
